+++ lib/extensions/logger.rs
@@ -133,7 +133,7 @@ impl Logger {
 
                 // Format Query as string
                 let mut query = "".to_string();
-                for i in &req.query.data {
+                for i in &req.query.0 {
                     query += &format!("{}: {}, ", i[0], i[1]);
                 }
                 if query.len() >= 2 {